+++ src/Basic/index.js
@@ -1,0 +1,4 @@
+// export { default as Avatar } from "./Avatar";
+export { default as Background } from "./Background";
+// export { default as Logo } from "./Logo";
+// export { default as Typography } from "./Typography";
