+++ src/Components/Comment/index.js
@@ -1,1 +1,1 @@
-export { default as CommentInput } from "./CommentInput";
+export { default } from "./CommentInput";
