+++ src/Components/Input/index.js
@@ -1,1 +1,1 @@
-export { default as Input } from "./Input";
+export { default } from "./Input";
