+++ src/Components/NavItem/index.js
@@ -1,1 +1,1 @@
-export { default as NavItem } from "./NavItem";
+export { default } from "./NavItem";
