+++ src/Components/Toggle/index.jsx
@@ -1,1 +1,1 @@
-export { default as Toggle } from "./Toggle";
+export { default } from "./Toggle";
