+++ src/Layouts/Navbar/Navbar.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import NavItem from "../../Components/NavItem/NavItem";
+import { NavItem } from "../../Components";
 import Logo from "../../Basic/Logo/Logo";
 import Icon from "../../Basic/Icon/Icon";
 import Avatar from "../../Basic/Avatar/Avatar";
