+++ src/Layouts/Navbar/index.js
@@ -1,1 +1,1 @@
-export { default as Navbar } from "./Navbar";
+export { default } from "./Navbar";
