+++ src/Layouts/Toc/index.js
@@ -1,1 +1,1 @@
-export { default as Toc } from "./Toc";
+export { default } from "./Toc";
