+++ src/Layouts/TocNav/TocNav.stories.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import TocNav from "./TocNav";
+import TocNav from "./";
 
 export default {
   title: "Layouts/TocNavbar",
